Create the whorl executor
This commit contains all of the work done in creating an initial version
such that anyone can learn how asynchronous executors in Rust work. It's
not fast by any means but it does work. The goal was focusing on
commenting on the code and making it all work such that someone could
read it beginning to end and get an idea of how all the pieces fit
together and play with it to their hearts content in a more interactive
way if they cloned the repo.