Change AtomicBool to bool
In previous iterations of whorl we needed an AtomicBool, but since we
wrap a Task in an Arc we don't actually need it since we never change
the value either. The Task either blocks or does not and that does not
change at all. We change to a normal bool and remove any phrasing
related to it being an AtomicBool.

Closes #4